+++ Keysential/Core/Commands/StartKeyManagerCommand.cs
@@ -1,5 +1,8 @@
 ﻿namespace Keysential;
 
+using System.Collections;
+using System.Collections.Generic;
+
 using ComfyLib;
 
 using UnityEngine;
@@ -8,45 +11,52 @@ public static class StartKeyManagerCommand {
   [ComfyCommand]
   public static Terminal.ConsoleCommand Register() {
     return new Terminal.ConsoleCommand(
-        "startkeymanager",
-        "startkeymanager <id: id1> <position: x,y,z> <distance: 8f> <keys: key1,key2> [add: true/false]",
-        args => Run(args));
+        "start-key-manager",
+        "(Keysential) start-key-manager --id=<manager-id> --position=<x,y,z> --distance=<8> --keys=<key1,key2> "
+            + "[--type=<Vendor|DistanceXZ>] [--add]",
+        Run);
   }
 
-  public static bool Run(Terminal.ConsoleEventArgs args) {
-    if (args.Length < 5) {
-      Keysential.LogError($"Not enough args for startkeymanager command.");
+  public static object Run(Terminal.ConsoleEventArgs args) {
+    return Run(new ComfyArgs(args));
+  }
+
+  public static bool Run(ComfyArgs args) {    if (!args.TryGetValue("id", out string managerId)) {
+      Keysential.LogError($"Missing string value for: --id");
       return false;
     }
 
-    string managerId = args[1];
 
-    if (!args[2].TryParseVector3(out Vector3 position)) {
-      Keysential.LogError($"Could nor parse Vector3 position arg: {args[2]}");
+    if (!args.TryGetValue("position", out Vector3 position)) {
+      Keysential.LogError($"Missing or invalid Vector3 value for: --position");
       return false;
     }
 
-    if (!float.TryParse(args[3], out float distance) || distance < 0f) {
-      Keysential.LogError($"Could not parse or invalid float distance arg: {args[3]}");
+    if (!args.TryGetValue("distance", out float distance) || distance < 0f) {
+      Keysential.LogError($"Missing or invalid float value for: --distance");
       return false;
     }
 
-    string[] keys = args[4].GetEncodedGlobalKeys();
-
-    if (keys.Length <= 0) {
-      Keysential.LogError($"No valid values for keys arg: {args[4]}");
+    if (!args.TryGetListValue("keys", out List<string> keys) || keys.Count <= 0) {
+      Keysential.LogError($"Missing or invalid string-list values for: --keys");
       return false;
     }
 
-    bool result =
-        GlobalKeysManager.StartKeyManager(
-            managerId,
-            position,
-            distance,
-            DistanceKeyManager.DistanceXZProximityCoroutine(managerId, position, distance, keys));
+    if (!args.TryGetValue("type", out GlobalKeysManager.ManagerType managerType)) {
+      managerType = GlobalKeysManager.ManagerType.DistanceXZ;
+    }
 
-    if (result && args.Length >= 6 && bool.TryParse(args[5], out bool addToStartUp) && addToStartUp) {
-      KeyManagerUtils.AddStartUpKeyManager($"startkeymanager {args[1]} {args[2]} {args[3]} {args[4]}");
+    if (!args.TryGetValue("add", out bool addToStartUp)) {
+      addToStartUp = false;
+    }
+
+    IEnumerator keyManagerCoroutine =
+        GlobalKeysManager.CreateKeyManagerCoroutine(managerId, managerType, position, distance, keys);
+
+    bool result = GlobalKeysManager.StartKeyManager(managerId, position, distance, keyManagerCoroutine);
+
+    if (result && addToStartUp) {
+      KeyManagerUtils.AddStartUpKeyManager(args.Args.FullLine);
     }
 
     return result;