+++ Keysential/Core/Commands/StopKeyManagerCommand.cs
@@ -6,21 +6,29 @@ public static class StopKeyManagerCommand {
   [ComfyCommand]
   public static Terminal.ConsoleCommand Register() {
     return new Terminal.ConsoleCommand(
-        "stopkeymanager",
-        "stopkeymanager <id: id1> [remove: true/false]",
-        args => Run(args));
+        "stop-key-manager",
+        "stop-key-manager --id=<manager-id> [--remove]",
+        Run);
   }
 
-  public static bool Run(Terminal.ConsoleEventArgs args) {
-    if (args.Length < 2) {
-      Keysential.LogError($"Not enough args for stopkeymanager command.");
+  public static object Run(Terminal.ConsoleEventArgs args) {
+    return Run(new ComfyArgs(args));
+  }
+
+  public static bool Run(ComfyArgs args) {
+    if (!args.TryGetValue("id", out string managerId)) {
+      Keysential.LogError($"Missing string value for: --id");
       return false;
     }
 
-    if (args.Length >= 3 && bool.TryParse(args[2], out bool removeFromStartup) && removeFromStartup) {
-      KeyManagerUtils.RemoveStartUpKeyManager(args[1]);
+    if (!args.TryGetValue("remove", out bool removeFromStartUp)) {
+      removeFromStartUp = false;
     }
 
-    return GlobalKeysManager.StopKeyManager(args[1]);
+    if (removeFromStartUp) {
+      KeyManagerUtils.RemoveStartUpKeyManager(managerId);
+    }
+
+    return GlobalKeysManager.StopKeyManager(managerId);
   }
 }